Add README with plugin overview
