+++ src/main/java/au/com/addstar/attributioner/Attributioner.java
@@ -25,6 +25,7 @@ public class Attributioner extends JavaPlugin implements Listener {
         AttributeManager manager = new AttributeManager(this);
         getServer().getPluginManager().registerEvents(new RegionListener(this, manager), this);
         getServer().getPluginManager().registerEvents(this, this);
+        getCommand("attributioner").setExecutor(new AttributionerCommand(this, manager));
     }
 
     public void loadConfig() {
@@ -46,7 +47,7 @@ public class Attributioner extends JavaPlugin implements Listener {
                     String opStr = section.getString(attrName + ".operation");
                     AttributeModifier.Operation op = AttributeModifier.Operation.valueOf(opStr);
 
-                    NamespacedKey modKey = new NamespacedKey("attributioner", attrName.toLowerCase());
+                    NamespacedKey modKey = new NamespacedKey("attributioner-" + regionName.toLowerCase(), attrName.toLowerCase());
                     //"Attributioner-" + regionName + "-" + attrName,
                     AttributeModifier modifier = new AttributeModifier(
                             modKey,
@@ -84,7 +85,7 @@ public class Attributioner extends JavaPlugin implements Listener {
             AttributeInstance instance = player.getAttribute(attribute);
             if (instance == null) continue;
             for (AttributeModifier modifier : instance.getModifiers()) {
-                if (modifier.getKey() != null && modifier.getKey().getNamespace().equals("attributioner")) {
+                if (modifier.getKey() != null && modifier.getKey().getNamespace().startsWith("attributioner-")) {
                     instance.removeModifier(modifier);
                 }
             }
